Refactor project column components and update switch styling
- Move project column and count components to shared directory
- Update import paths in collections and projects modules
- Change switch checked background color to primary-500
- Improve template rendering in collection page
