+++ frontend/config/styles/components/switch.scss
@@ -19,7 +19,7 @@
   --lf-switch-handle: var(--lf-color-white);
 
   // Checked
-  --lf-switch-checked-background: var(--lf-color-gray-900);
+  --lf-switch-checked-background: var(--lf-color-primary-500);
   --lf-switch-checked-handle: var(--lf-color-white);
 
   // Disabled Unchecked
