+++ frontend/src/shared/project-column/lf-project-count.vue
@@ -1,8 +1,8 @@
 <template>
   <div>
     <lf-badge type="secondary" class="!flex items-center gap-1 w-full !rounded-full !px-2.5">
-      <lf-icon name="layer-group" type="regular" />
-      {{ pluralize('project', props.count, true) }}
+      <lf-icon :name="props.icon || 'layer-group'" type="regular" />
+      {{ pluralize(props.title || 'project', props.count, true) }}
     </lf-badge>
   </div>
 </template>
@@ -14,6 +14,8 @@ import LfBadge from '@/ui-kit/badge/Badge.vue';
 
 const props = defineProps<{
   count: number;
+  icon?: string;
+  title?: string;
 }>();
 </script>
